motion_sensor_project/requirements.txt
# Python libraries needed for this project
# These can be installed using pip: pip3 install -r requirements.txt

RPi.GPIO>=0.7.0
numpy>=1.19.0
pygame>=2.0.0
